LayoutTests/imported/w3c/web-platform-tests/intersection-observer/inline-client-rect-expected.txt

1 2 3 4 5

PASS Inline target 
FAIL First rAF assert_equals: entries[0].boundingClientRect.left expected 320 but got 0
FAIL scroller.scrollLeft = 90 assert_equals: entries.length expected 2 but got 1
